Final commit of orix content dump
